Incorporate Wadim's review comments
