+++ package.json
@@ -71,7 +71,7 @@
     },
     "dependencies": {
         "@formatjs/intl": "^2.6.3",
-        "i18nlint-common": "^1.4.0",
+        "i18nlint-common": "^1.3.0",
         "ilib-common": "^1.1.3",
         "ilib-locale": "^1.2.2",
         "ilib-localeinfo": "^1.0.5",
